Updates to refactor mail services and add template mail service
