+++ src/dto/process-mail.dto.ts
@@ -1,0 +1,8 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsNumber } from 'class-validator';
+
+export class ProcessMailDTO {
+  @ApiProperty()
+  @IsNumber()
+  emailToProcessId: number;
+}
