+++ src/mail/mail.controller.spec.ts
@@ -2,19 +2,33 @@ import { HttpModule } from '@nestjs/axios';
 import { ConfigService } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { LoggerModule } from '@us-epa-camd/easey-common/logger';
-import { MassEvalParamsDTO } from '../dto/mass-eval-params.dto';
 import { CreateMailDto } from '../dto/create-mail.dto';
 import { MailController } from './mail.controller';
 import { MailService } from './mail.service';
+import { ProcessMailDTO } from '../dto/process-mail.dto';
+import { MailEvalService } from './mail-eval.service';
+import { MailTemplateService } from './mail-template.service';
+import { MassEvalParamsDTO } from '../dto/mass-eval-params.dto';
 
 const mockMailService = () => ({
   sendEmail: jest.fn(),
   sendMassEvalEmail: jest.fn(),
+  sendEmailRecord: jest.fn(),
+});
+
+const mockEvalService = () => ({
+  sendMassEvalEmail: jest.fn(),
+});
+
+const mockTemplateService = () => ({
+  sendEmailRecord: jest.fn(),
 });
 
 describe('Mail Controller', () => {
   let controller: MailController;
   let service: MailService;
+  let evalService: MailEvalService;
+  let templateService: MailTemplateService;
 
   beforeAll(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -22,11 +36,15 @@ describe('Mail Controller', () => {
       controllers: [MailController],
       providers: [
         { provide: MailService, useFactory: mockMailService },
+        { provide: MailEvalService, useFactory: mockEvalService },
+        { provide: MailTemplateService, useFactory: mockTemplateService },
         ConfigService,
       ],
     }).compile();
 
     service = module.get(MailService);
+    evalService = module.get(MailEvalService);
+    templateService = module.get(MailTemplateService);
     controller = module.get(MailController);
   });
 
@@ -40,9 +58,15 @@ describe('Mail Controller', () => {
     expect(service.sendEmail).toHaveBeenCalled();
   });
 
-  it('should call the mass eval service', () => {
+  it('should call the mass-eval service', () => {
     controller.sendMassEval(new MassEvalParamsDTO());
 
-    expect(service.sendMassEvalEmail).toHaveBeenCalled();
+    expect(evalService.sendMassEvalEmail).toHaveBeenCalled();
+  });
+
+  it('should call the template service', () => {
+    controller.sendRecord(new ProcessMailDTO());
+
+    expect(templateService.sendEmailRecord).toHaveBeenCalled();
   });
 });
